test(infra): share auth/token test doubles via src/testing, migrate Authorization
Lift the JWT-cookie and access-token test doubles that were duplicated across
the AccessManagement and Authorization apps into a single shared source-linked
set under src/testing/AuthTestInfra, in the existing Altinn.Authorization.Testing
namespace: TestSigningCertificate (one in-memory self-signed cert), plus the
ConfigurationManagerStub, JwtCookiePostConfigureOptionsStub,
PublicSigningKeyProviderMock and JwtTokenMock that pair with it.

Opt-in via a new IncludeAuthTestInfra MSBuild property (mirrors the existing
IncludePostgresTestEngine pattern), enabled here for Altinn.Authorization.Tests.
Authorization already validated tokens against a single in-memory certificate,
so this migration is behaviour-preserving: its five local copies are removed and
consumers resolve the shared types through the global using.

AccessManagement keeps its own copies for now; it carries a per-issuer org-cert
path (platform-org/ttd-org) and a separate signing-key resolver that need their
own verified migration.

Verified: Altinn.Authorization.Tests builds clean and all 415 tests pass under
Podman.
