+++ src/apps/Altinn.Authorization/test/Altinn.Authorization.Tests/Altinn.Authorization.Tests.csproj
@@ -7,6 +7,8 @@
         <TargetFrameworks>net10.0</TargetFrameworks>
         <!-- Link the shared PostgreSQL test engine (src/testing/PostgresTestEngine.cs). -->
         <IncludePostgresTestEngine>true</IncludePostgresTestEngine>
+        <!-- Link the shared auth/token test doubles (src/testing/AuthTestInfra/*.cs). -->
+        <IncludeAuthTestInfra>true</IncludeAuthTestInfra>
     </PropertyGroup>
 
     <ItemGroup>
