--- src/apps/Altinn.Authorization/test/Altinn.Authorization.Tests/MockServices/JwtCookiePostConfigureOptionsStub.cs
@@ -1,40 +1,0 @@
-using System;
-
-using AltinnCore.Authentication.JwtCookie;
-
-using Microsoft.AspNetCore.Authentication.Cookies;
-using Microsoft.Extensions.Options;
-
-namespace Altinn.Authorization.Tests.MockServices
-{
-    /// <summary>
-    /// Represents a stub for the <see cref="JwtCookiePostConfigureOptions"/> class to be used in integration tests.
-    /// </summary>
-    public class JwtCookiePostConfigureOptionsStub : IPostConfigureOptions<JwtCookieOptions>
-    {
-        /// <inheritdoc />
-        public void PostConfigure(string name, JwtCookieOptions options)
-        {
-            if (string.IsNullOrEmpty(options.JwtCookieName))
-            {
-                options.JwtCookieName = JwtCookieDefaults.CookiePrefix + name;
-            }
-
-            if (options.CookieManager == null)
-            {
-                options.CookieManager = new ChunkingCookieManager();
-            }
-
-            if (!string.IsNullOrEmpty(options.MetadataAddress))
-            {
-                if (!options.MetadataAddress.EndsWith("/", StringComparison.Ordinal))
-                {
-                    options.MetadataAddress += "/";
-                }
-            }
-
-            options.MetadataAddress += ".well-known/openid-configuration";
-            options.ConfigurationManager = new ConfigurationManagerStub();
-        }
-    }
-}
