--- src/apps/Altinn.Authorization/test/Altinn.Authorization.Tests/Util/JwtTokenMock.cs
@@ -1,39 +1,0 @@
-﻿using System;
-using System.IdentityModel.Tokens.Jwt;
-using System.Security.Claims;
-
-using Microsoft.IdentityModel.Tokens;
-
-namespace Altinn.Authorization.Tests.Util
-{
-    /// <summary>
-    /// Represents a mechanism for creating JSON Web tokens for use in integration tests.
-    /// </summary>
-    public static class JwtTokenMock
-    {
-        /// <summary>
-        /// Generates a token with a self signed certificate included in the integration test project.
-        /// </summary>
-        /// <param name="principal">The claims principal to include in the token.</param>
-        /// <param name="tokenExpiry">How long the token should be valid for.</param>
-        /// <param name="issuer">The URL of the token issuer</param>
-        /// <returns>A new token.</returns>
-        public static string GenerateToken(ClaimsPrincipal principal, TimeSpan tokenExpiry, string issuer = "UnitTest")
-        {
-            JwtSecurityTokenHandler tokenHandler = new JwtSecurityTokenHandler();
-            SecurityTokenDescriptor tokenDescriptor = new SecurityTokenDescriptor
-            {
-                Subject = new ClaimsIdentity(principal.Identity),
-                Expires = DateTime.UtcNow.AddSeconds(tokenExpiry.TotalSeconds),
-                SigningCredentials = TestCertificates.SigningCredentials,
-                Audience = "altinn.no",
-                Issuer = issuer
-            };
-
-            SecurityToken token = tokenHandler.CreateToken(tokenDescriptor);
-            string serializedToken = tokenHandler.WriteToken(token);
-
-            return serializedToken;
-        }
-    }
-}
